Refactored controller class
