+++ userservice/src/main/java/com/perficient/userservice/controller/UserCreateController.java
@@ -1,0 +1,22 @@
+package com.perficient.userservice.controller;
+
+import com.perficient.userservice.dto.UserDto;
+import com.perficient.userservice.service.UserCreateService;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/v1")
+@RestController
+public class UserCreateController {
+
+    @Autowired
+    private UserCreateService createService;
+    @PostMapping("/user")
+    public UserDto createUser(@Valid @RequestBody UserDto user) {
+        return createService.createUser(user);
+    }
+}
